feat: Upgrade Go version in Dockerfile and refactor type imports for better organization
