+++ client/src/types/query.ts
